requirements.txt
requests
beautifulsoup4
python-telegram-bot==20.7
phonenumbers
country_converter